+++ src/components/App/App.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { ThemeProvider } from "@material-ui/core/styles";
-import { theme } from "./theme";
+import { theme } from "../../theme";
 import "./App.css";
-import { Routes } from "./Routes";
+import { Routes } from "../../routing/Routes";
 
 const App = () => {
   return (
