+++ src/components/Header/Header.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Grid, AppBar, Toolbar } from "@material-ui/core";
-import { ReactComponent as LogoIcon } from "./logoicon.svg";
-import { ReactComponent as LogoText } from "./lofttaxi.svg";
-import Nav from "./Nav";
+import { ReactComponent as LogoIcon } from "../../img/logoicon.svg";
+import { ReactComponent as LogoText } from "../../img/lofttaxi.svg";
+import Nav from "../Nav/Nav";
 
 export const Header = () => {
   return (
