+++ src/components/Login/Login.spec.jsx
@@ -1,6 +1,6 @@
 import Login from "./Login";
 import { render } from "@testing-library/react";
-import { TestWrapper } from "./testWrapper";
+import { TestWrapper } from "../../testWrapper";
 
 describe("Login", () => {
   it("renders correctly", () => {
