+++ src/components/Register/Register.spec.jsx
@@ -1,7 +1,6 @@
 import Register from "./Register";
 import { render } from "@testing-library/react";
-import { MemoryRouter as Router } from "react-router-dom";
-import { TestWrapper } from "./testWrapper";
+import { TestWrapper } from "../../testWrapper";
 
 describe("Register", () => {
   it("renders correctly", () => {
