+++ src/components/Welcome/Welcome.jsx
@@ -1,10 +1,10 @@
 import React from "react";
-import Login from "./Login";
-import Register from "./Register";
-import { ReactComponent as Logo } from "./logo.svg";
+import Login from "../Login/Login";
+import Register from "../Register/Register";
+import { ReactComponent as Logo } from "../../img/logo.svg";
 import { Grid } from "@material-ui/core";
 import { PropTypes } from "prop-types";
-import { loginStateSelector } from "./reducers/rootReducer";
+import { loginStateSelector } from "../../reducers/rootReducer";
 import { Redirect } from "react-router-dom";
 import { connect } from "react-redux";
