+++ src/routing/PrivateRoute.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route, Redirect } from "react-router-dom";
 import { connect } from "react-redux";
-import { loginStateSelector } from "./reducers/rootReducer";
+import { loginStateSelector } from "../reducers/rootReducer";
 
 const PrivateRoute = ({ children, ...props }) => {
   console.log(props);
